requirements.txt
cryptography==36.0.1
flask>=2.0.2
Flask-SQLAlchemy==2.4.3
flask-restx==0.5.1
python-dotenv==0.19.2
PyJWT==2.3.0
